refactor: change in structure
